refactor: unify behaviours in a single file
