+++ peak/behaviours.py
@@ -1,35 +1,109 @@
 import logging
+from abc import ABCMeta, abstractmethod
 from json import dumps as json_dumps
 from typing import Callable
 
+import aioxmpp
+import spade.behaviour
+from aioxmpp import JID
+from spade.message import Message
+
 import peak
+from peak import Agent
 from peak.df import jid as df_jid
 
-
-class CreateNode(peak.OneShotBehaviour):
-    def __init__(self, node: str, affiliation: str = None):
-        super().__init__()
-        self.affiliation = affiliation
-        self.node = node
-
-    async def on_start(self):
-        self.template = peak.Template()
-        self.template.set_metadata("resource", "pubsub_create_node")
-
-    async def run(self):
-        msg = peak.Message()
-        msg.to = df_jid(self.agent.jid.domain)
-        msg.set_metadata("resource", "pubsub_create_node")
-        msg.set_metadata("affiliation", self.affiliation if self.affiliation else "")
-        msg.set_metadata("node_jid", self.node)
-        await self.send(msg)
-        res = await self.receive(60)
-        if res:
-            logger = logging.getLogger("CreateNode")
-            logger.info("PubSub node " + self.node + " created")
+aioxmpp.pubsub.xso.as_payload_class(aioxmpp.Message)
 
 
-class JoinGroup(peak.OneShotBehaviour):
+def slice_jid(jid):
+    jid = JID.fromstr(jid)
+    pubsub = JID.fromstr(jid.domain)
+    node = jid.localpart
+    return pubsub, node
+
+
+class _Behaviour:
+
+    agent: Agent
+
+    async def send_to_group(self, msg: Message):
+        """Sends a message to a group chat.
+
+        When sending a message to a group the agent joins the group first. The parameter
+        'leave' tells the method if the agent leaves, or not, the group after sending the
+        message. (If the intention is to send a single request to a new group the best
+        option would be to leave the group chat, if the intention is to send a message
+        to a group wich the agent already belongs to, it's better to not leave)
+        Args:
+            msg (mas.Message): The Message.
+            group (str, optional): Name of the group to send the message to. If None is given the
+                                   the message is sent to the MAS group. Defaults to None.
+            leave (bool, optional): If true, agent leaves the group after sending the message. Defaults to False.
+        """
+        raw_msg = msg.prepare()
+        try:
+            await self.agent.groups[str(msg.to)].send_message(raw_msg)
+        except:
+            room, future = self.agent.muc_client.join(msg.to, self.agent.name)
+            await future
+            await room.send_message(raw_msg)
+            await room.leave()
+
+    async def change_node_affiliations(self, jid: str, affiliations_to_set: tuple):
+        """Changes PubSub node affiliations.
+
+        Args:
+            jid (str): JID of the node, e.g. node@pubsub.example.com
+            affiliations_to_set (tuple[str,str]): each tuple must contain the JID of the user and the affiliation (e.g.'owner','publisher')
+        """
+        pubsub, node = slice_jid(jid)
+        user, aff = affiliations_to_set
+        await self.agent.pubsub_client.change_node_affiliations(
+            pubsub, node, [(JID.fromstr(user), aff)]
+        )
+
+    async def subscribe(self, jid: str):
+        pubsub, node = slice_jid(jid)
+        await self.agent.pubsub_client.subscribe(pubsub, node)
+        await self.agent.pubsub_client.on_item_published.connect(self.on_item_published)
+
+    async def publish(self, msg: Message):
+        jid = msg.to.domain
+        node = msg.to.localpart
+        await self.agent.pubsub_client.publish(JID.fromstr(jid), node, msg.prepare())
+
+    async def notify(self, jid: str):
+        pubsub, node = slice_jid(jid)
+        await self.agent.pubsub_client.notify(pubsub, node)
+
+    async def unsubscribe(self, jid: str):
+        pubsub, node = slice_jid(jid)
+        await self.agent.pubsub_client.unsubscribe(pubsub, node)
+
+    @abstractmethod
+    def on_item_published(jid, node, item, *, message=None):
+        pass
+
+
+class OneShotBehaviour(spade.behaviour.OneShotBehaviour, _Behaviour, metaclass=ABCMeta):
+    pass
+
+
+class PeriodicBehaviour(
+    spade.behaviour.PeriodicBehaviour, _Behaviour, metaclass=ABCMeta
+):
+    pass
+
+
+class CyclicBehaviour(spade.behaviour.CyclicBehaviour, _Behaviour, metaclass=ABCMeta):
+    pass
+
+
+class FSMBehaviour(spade.behaviour.FSMBehaviour, _Behaviour, metaclass=ABCMeta):
+    pass
+
+
+class JoinGroup(OneShotBehaviour):
     def __init__(self, path: str, domain: str, tags: list = []):
         super().__init__()
         self.path = path
@@ -51,7 +125,7 @@ class JoinGroup(peak.OneShotBehaviour):
         await self.agent.join_group(nodes[-1] + "_down@" + self.domain)
 
 
-class LeaveGroup(peak.OneShotBehaviour):
+class LeaveGroup(OneShotBehaviour):
     def __init__(self, path: str, domain: str):
         super().__init__()
         self.path = path
@@ -72,7 +146,7 @@ class LeaveGroup(peak.OneShotBehaviour):
         await self.agent.leave_group(nodes[-1] + "_down@" + self.domain)
 
 
-class SearchGroup(peak.OneShotBehaviour):
+class SearchGroup(OneShotBehaviour):
     def __init__(
         self, tags: list[str], callback: Callable[[list[str]], None], *args, **kargs
     ):
@@ -97,7 +171,7 @@ class SearchGroup(peak.OneShotBehaviour):
             self.callback(self.tags, groups, *self.args, **self.kargs)
 
 
-class ExportData(peak.OneShotBehaviour):
+class ExportData(OneShotBehaviour):
     def __init__(
         self,
         file_name: str,
@@ -144,7 +218,7 @@ class ExportData(peak.OneShotBehaviour):
             )
 
 
-class _ExportDataSync(peak.CyclicBehaviour):
+class _ExportDataSync(CyclicBehaviour):
     def __init__(
         self,
         file_name: str,
@@ -215,7 +289,7 @@ class _ExportDataSync(peak.CyclicBehaviour):
                     f.write(json_dumps(self.data))
 
 
-class _ExportData(peak.PeriodicBehaviour):
+class _ExportData(PeriodicBehaviour):
     def __init__(
         self,
         file_name: str,
